Add files via upload
Made it so resetting your turn does not pass it to the next player. Game should be operational in a basic form, now. I may add the rules for getting the King to a corner, hostile corners, and a hostile throne some time later, but for now the game is playable, with a few caveats.